ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.10.3
PyPDF2==3.0.1
openai==1.3.0
groq==0.4.0
gTTS==2.5.0
pydub==0.25.1
moviepy==1.0.3
Pillow==10.1.0
numpy==1.26.2
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.2